develop(Backend): correciones y mejores en las historias de usuario

+++ src/main/java/com/techcommunityperu/techcommunityperu/repository/UserRepository.java
@@ -3,7 +3,7 @@ package com.techcommunityperu.techcommunityperu.repository;
 import com.techcommunityperu.techcommunityperu.model.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Usuario, Long> {
+public interface UserRepository extends JpaRepository<Usuario, Integer> {
     boolean existsByCorreoElectronico(String correoElectronico);
     Usuario findByCorreoElectronico(String correoElectronico);
 }
